Move http package inside cmd
