+++ cmd/gateway-main.go
@@ -27,8 +27,8 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/minio/cli"
+	miniohttp "github.com/minio/minio/cmd/http"
 	"github.com/minio/minio/pkg/errors"
-	miniohttp "github.com/minio/minio/pkg/http"
 )
 
 var (
